+++ src/noise_math.h
@@ -3,7 +3,7 @@
 // Based on Ken Perlin's smoother step function:
 // https://en.wikipedia.org/wiki/Smoothstep#Variations Returns value in range
 // [0, 1]
-inline constexpr double SmootherStep(const double x) {
+inline constexpr float SmootherStep(const float x) {
   if (x < 0) {
     return 0;
   }
@@ -17,7 +17,7 @@ inline constexpr double SmootherStep(const double x) {
 // Interpolates between two doubles
 // If x <= 0 returns bound_0, if x >= 1 returns bound_1,
 // otherwise smoothly transitions between the two
-inline constexpr double Interpolate(const double x, const double bound_0,
-                                    const double bound_1) {
+inline constexpr float Interpolate(const float x, const float bound_0,
+                                   const float bound_1) {
   return bound_0 + SmootherStep(x) * (bound_1 - bound_0);
 }
